refactor(table): move local-ISO date formatters to utils/dates
dateToISODate/Time/DateTime were generic Date-to-string helpers stranded
in filters/serialize.ts after the folder split. Move them to utils/dates
alongside the existing date formatters, rename to make the local-time
semantics explicit (dateToLocalISO*), and add docstrings clarifying when
to use them versus Date.toISOString().
