+++ frontend/src/components/data-table/__tests__/date-filter-inputs.test.tsx
@@ -1,7 +1,7 @@
 /* Copyright 2026 Marimo. All rights reserved. */
 import { describe, expect, it } from "vitest";
+import { dateToLocalISODate } from "@/utils/dates";
 import { parsePastedRange } from "../date-filter-inputs";
-import { dateToISODate } from "../filters";
 
 describe("parsePastedRange", () => {
   it.each([
@@ -15,15 +15,15 @@ describe("parsePastedRange", () => {
   ])("splits a date range pasted with %s separator", (_, text) => {
     const result = parsePastedRange("date", text);
     expect(result).toBeDefined();
-    expect(result && dateToISODate(result.min)).toBe("2026-02-01");
-    expect(result && dateToISODate(result.max)).toBe("2026-04-01");
+    expect(result && dateToLocalISODate(result.min)).toBe("2026-02-01");
+    expect(result && dateToLocalISODate(result.max)).toBe("2026-04-01");
   });
 
   it("returns a degenerate range for a single pasted date", () => {
     const result = parsePastedRange("date", "2026-03-15");
     expect(result).toBeDefined();
-    expect(result && dateToISODate(result.min)).toBe("2026-03-15");
-    expect(result && dateToISODate(result.max)).toBe("2026-03-15");
+    expect(result && dateToLocalISODate(result.min)).toBe("2026-03-15");
+    expect(result && dateToLocalISODate(result.max)).toBe("2026-03-15");
   });
 
   it("returns undefined for unparsable input", () => {
